ActionPathList is now deprecated and will be removed in v5.0.0.
